+++ src/components/App/App.tsx
@@ -1,89 +1,52 @@
+/* eslint-disable react/no-unstable-nested-components */
 import React from 'react';
-import { Alert, CssBaseline, ThemeProvider } from '@mui/material';
-import useTheme from '@mui/material/styles/useTheme';
-import useMediaQuery from '@mui/material/useMediaQuery';
+import { CssBaseline, Paper, Box, Typography } from '@mui/material';
+import { ThemeProvider } from '@mui/material/styles';
 import { useTranslation } from 'react-i18next';
-import {
-  CopyrightModal,
-  Footer,
-  Header,
-  Spacer,
-  SPACER_VERTICAL
-} from '@ska-telescope/ska-gui-components';
 import { storageObject } from '@ska-telescope/ska-gui-local-storage';
-import Loader from '../Loader/Loader';
-import ReactSkeleton from '../ReactSkeleton/ReactSkeleton';
+import { UserProfile } from '../UserProfile/UserProfile';
 import theme from '../../services/theme/theme';
-
-const HEADER_HEIGHT = 70;
-const FOOTER_HEIGHT = 20;
+import { fullHeight } from '../../utils/constants';
+import Loader from '../Loader/Loader';
+import { MsalAuthenticationTemplate, UnauthenticatedTemplate } from '@azure/msal-react';
+import { InteractionType } from '@azure/msal-browser';
+import { Shell } from '../Shell/Shell';
 
 function App() {
-  const { t } = useTranslation('reactSkeleton');
-  const [showCopyright, setShowCopyright] = React.useState(false);
-  const { help, helpToggle, telescope, themeMode, toggleTheme, updateTelescope } =
-    storageObject.useStore();
-
-  const LG = () => useMediaQuery(useTheme().breakpoints.down('lg')); // Allows us to code depending upon screen size
-  const REQUIRED_WIDTH = useMediaQuery('(min-width:600px)');
-
-  const skao = t('toolTip.button.skao');
-  const mode = t('toolTip.button.mode');
-  const headerTip = t('toolTip.button.docs');
-  const headerURL = t('toolTip.button.docsURL');
-  const docs = { tooltip: headerTip, url: headerURL };
-  const toolTip = { skao, mode };
-  const version = process.env.VERSION;
-  const theStorage = {
-    help,
-    helpToggle,
-    telescope,
-    themeMode: themeMode.mode,
-    toggleTheme,
-    updateTelescope
-  };
-
-  const mediaSizeNotSupported = () => <Alert>{t('mediaSize.notSupported')}</Alert>;
+  const { themeMode } = storageObject.useStore();
+  const { t } = useTranslation('authentication');
 
   return (
-    <ThemeProvider theme={theme(themeMode.mode)}>
+    <ThemeProvider theme={theme(themeMode?.mode)}>
       <CssBaseline enableColorScheme />
       <React.Suspense fallback={<Loader />}>
-        {
-          // Header container :
-          // Even distribution of the children is built in
-          // Logo with URL link included
-          // Button for light/dark mode included, and sample implementation provided.
-          // TelescopeSelector build in, displayed as determined by selectTelescope property
-        }
-        <CopyrightModal copyrightFunc={setShowCopyright} show={showCopyright} />
-        <Header
-          docs={docs}
-          testId="headerId"
-          title={LG() ? 'SRK' : 'SKA React Skeleton'} // Use a 3 letter code for smaller screen widths
-          toolTip={toolTip}
-          selectTelescope={false}
-          storage={theStorage}
-          useSymbol={LG()}
-        />
-        {
-          // Example of the spacer being used to shift content from behind the Header component
-        }
-        <Spacer size={HEADER_HEIGHT} axis={SPACER_VERTICAL} />
-        {
-          // This is the ONLY component that is accessible via micro-frontend implementation
-        }
-        {REQUIRED_WIDTH && <ReactSkeleton />}
-        {!REQUIRED_WIDTH && mediaSizeNotSupported()}
-        {
-          // Example of the spacer being used to stop content from being hidden behind the Footer component
-        }
-        <Spacer size={FOOTER_HEIGHT} axis={SPACER_VERTICAL} />
-        {
-          // Footer container :
-          // Even distribution of the children is built in
-        }
-        <Footer copyrightFunc={setShowCopyright} testId="footerId" version={version} />
+        <Paper sx={{ height: '100vh' }}>
+          <Shell>
+            <Box
+              m={5}
+              sx={{
+                height: fullHeight(),
+                display: 'flex',
+                flexDirection: 'column',
+                justifyContent: 'center',
+                alignItems: 'center'
+              }}
+            >
+              <Typography component="h1" variant="h5">
+                {t('label.usingMSEntra')}
+              </Typography>
+              <Typography component="h1" variant="h5">
+                <MsalAuthenticationTemplate interactionType={InteractionType.None}>
+                  <p>{t('label.msEntraUserSignedIn')}</p>
+                </MsalAuthenticationTemplate>
+                <UnauthenticatedTemplate>
+                  <p>{t('label.msEntraUserNotSignedIn')}</p>
+                </UnauthenticatedTemplate>
+              </Typography>
+              <UserProfile />
+            </Box>
+          </Shell>
+        </Paper>
       </React.Suspense>
     </ThemeProvider>
   );