+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Changelog
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+1.1.0
+
+Addition of the ska-login components as a working example
+
 1.0.3
 
 ---