+++ docs/src/CHANGELOG.rst
@@ -4,6 +4,10 @@ Changelog
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+1.1.0
+
+Addition of the ska-login library to provide an example of the login/logout
+
 1.0.3
 
 ---